dummy-data-product/src/requirements.txt
python3
request 
bs4
pandas
os 
sys